requirements.txt
streamlit
pandas
numpy
requests
Pillow
opencv-python
python-dotenv
openpyxl